Add layout structure
