+++ layouts/footer.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export default () => (
+  <div className="footer">
+    <span>Also find Josh on </span>
+    <a href="https://github.com/hawkins">GitHub</a>
+    <span> or </span>
+    <a href="https://twitter.com/hawkinjs">Twitter</a>
+    <span>!</span>
+  </div>
+)
